REFACTOR: SimulationSetup part 1 - Move all classes
