+++ src/pyedb/dotnet/edb.py
@@ -51,7 +51,7 @@ from pyedb.dotnet.edb_core.edb_data.edbvalue import EdbValue
 from pyedb.dotnet.edb_core.edb_data.hfss_pi_simulation_setup_data import (
     HFSSPISimulationSetup,
 )
-from pyedb.dotnet.edb_core.edb_data.hfss_simulation_setup_data import (
+from pyedb.dotnet.edb_core.sim_setup_data.data.simulation_setup import (
     HfssSimulationSetup,
 )
 from pyedb.dotnet.edb_core.edb_data.ports import (
