+++ src/pyedb/dotnet/edb_core/edb_data/hfss_pi_simulation_setup_data.py
@@ -22,14 +22,14 @@
 
 from pyedb.dotnet.edb_core.general import convert_py_list_to_net_list
 from pyedb.dotnet.edb_core.utilities.simulation_setup import (
-    BaseSimulationSetup,
     EdbFrequencySweep,
 )
+from pyedb.dotnet.edb_core.sim_setup_data.data.simulation_setup import SimulationSetup
 from pyedb.generic.data_handlers import pyedb_function_handler
 from pyedb.generic.general_methods import generate_unique_name
 
 
-class HFSSPISimulationSetup(BaseSimulationSetup):
+class HFSSPISimulationSetup(SimulationSetup):
     """Manages EDB methods for HFSSPI simulation setup."""
 
     def __init__(self, pedb, edb_object=None):
