+++ src/pyedb/dotnet/edb_core/sim_setup_data/data/adaptive_frequency_data.py
@@ -1,0 +1,72 @@
+# Copyright (C) 2023 - 2024 ANSYS, Inc. and/or its affiliates.
+# SPDX-License-Identifier: MIT
+#
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+
+class AdaptiveFrequencyData(object):
+    """Manages EDB methods for adaptive frequency data."""
+
+    def __init__(self, adaptive_frequency_data):
+        self._adaptive_frequency_data = adaptive_frequency_data
+
+    @property
+    def adaptive_frequency(self):
+        """Adaptive frequency for the setup.
+
+        Returns
+        -------
+        str
+            Frequency with units.
+        """
+        return self._adaptive_frequency_data.AdaptiveFrequency
+
+    @adaptive_frequency.setter
+    def adaptive_frequency(self, value):
+        self._adaptive_frequency_data.AdaptiveFrequency = value
+
+    @property
+    def max_delta(self):
+        """Maximum change of S-parameters between two consecutive passes, which serves as
+        a stopping criterion.
+
+        Returns
+        -------
+        str
+        """
+        return self._adaptive_frequency_data.MaxDelta
+
+    @max_delta.setter
+    def max_delta(self, value):
+        self._adaptive_frequency_data.MaxDelta = str(value)
+
+    @property
+    def max_passes(self):
+        """Maximum allowed number of mesh refinement cycles.
+
+        Returns
+        -------
+        int
+        """
+        return self._adaptive_frequency_data.MaxPasses
+
+    @max_passes.setter
+    def max_passes(self, value):
+        self._adaptive_frequency_data.MaxPasses = value
