+++ src/pyedb/dotnet/edb_core/utilities/simulation_setup.py
@@ -23,253 +23,6 @@
 from pyedb.generic.general_methods import generate_unique_name, pyedb_function_handler
 
 
-class BaseSimulationSetup(object):
-    """Provide base simulation setup.
-
-    Parameters
-    ----------
-    pedb : :class:`pyedb.dotnet.edb.Edb`
-        Inherited object.
-    edb_object : :class:`Ansys.Ansoft.Edb.Utility.SIWaveSimulationSetup`,
-    :class:`Ansys.Ansoft.Edb.Utility.SIWDCIRSimulationSettings`,
-    :class:`Ansys.Ansoft.Edb.Utility.HFSSSimulationSettings`
-        EDB object.
-    """
-
-    @pyedb_function_handler
-    def __init__(self, pedb, edb_setup=None):
-        self._pedb = pedb
-        self._edb_object = edb_setup
-        self._setup_type = ""
-        self._setup_type_mapping = {
-            "kHFSS": self._pedb.simsetupdata.HFSSSimulationSettings,
-            "kPEM": None,
-            "kSIwave": self._pedb.simsetupdata.SIwave.SIWSimulationSettings,
-            "kLNA": None,
-            "kTransient": None,
-            "kQEye": None,
-            "kVEye": None,
-            "kAMI": None,
-            "kAnalysisOption": None,
-            "kSIwaveDCIR": self._pedb.simsetupdata.SIwave.SIWDCIRSimulationSettings,
-            "kSIwaveEMI": None,
-            "kHFSSPI": None,
-            "kDDRwizard": None,
-            "kQ3D": None,
-            "kNumSetupTypes": None,
-        }
-
-        version = self._pedb.edbversion.split(".")
-        if float(self._pedb.edbversion) >= 2024.2:
-            self._setup_type_mapping.update(
-                {
-                    "kRaptorX": self._pedb.simsetupdata.RaptorX.RaptorXSimulationSettings,
-                    "kHFSSPI": self._pedb.simsetupdata.HFSSPISimulationSettings,
-                }
-            )
-        if self._edb_object:
-            self._name = self._edb_object.GetName()
-
-        self._sweep_list = {}
-
-    @pyedb_function_handler
-    def _create(self, name=None):
-        """Create a simulation setup."""
-        if not name:
-            name = generate_unique_name(self.setup_type)
-            self._name = name
-
-        setup_type = self._setup_type_mapping[self._setup_type]
-        edb_setup_info = self._pedb.simsetupdata.SimSetupInfo[setup_type]()
-        edb_setup_info.Name = name
-        if (
-            edb_setup_info.get_SimSetupType().ToString() == "kRaptorX"
-            or edb_setup_info.get_SimSetupType().ToString() == "kHFSSPI"
-        ):
-            self._edb_setup_info = edb_setup_info
-        self._edb_object = self._set_edb_setup_info(edb_setup_info)
-        self._update_setup()
-
-    @pyedb_function_handler
-    def _set_edb_setup_info(self, edb_setup_info):
-        """Create a setup object from a setup information object."""
-        utility = self._pedb._edb.Utility
-        setup_type_mapping = {
-            "kHFSS": utility.HFSSSimulationSetup,
-            "kPEM": None,
-            "kSIwave": utility.SIWaveSimulationSetup,
-            "kLNA": None,
-            "kTransient": None,
-            "kQEye": None,
-            "kVEye": None,
-            "kAMI": None,
-            "kAnalysisOption": None,
-            "kSIwaveDCIR": utility.SIWaveDCIRSimulationSetup,
-            "kSIwaveEMI": None,
-            "kHFSSPI": None,
-            "kDDRwizard": None,
-            "kQ3D": None,
-            "kNumSetupTypes": None,
-        }
-
-        version = self._pedb.edbversion.split(".")
-        if int(version[0]) == 2024 and int(version[1]) == 2 or int(version[0]) > 2024:
-            setup_type_mapping["kRaptorX"] = utility.RaptorXSimulationSetup
-        setup_utility = setup_type_mapping[self._setup_type]
-        return setup_utility(edb_setup_info)
-
-    @pyedb_function_handler()
-    def _update_setup(self):
-        """Update setup in EDB."""
-        if self._setup_type == "kHFSS":
-            mesh_operations = self.get_sim_setup_info.SimulationSettings.MeshOperations
-            mesh_operations.Clear()
-            for mop in self.mesh_operations.values():
-                mesh_operations.Add(mop.mesh_operation)
-
-        if self._name in self._pedb.setups:
-            self._pedb.layout.cell.DeleteSimulationSetup(self._name)
-        if not self._pedb.layout.cell.AddSimulationSetup(self._edb_object):
-            raise Exception("Updating setup {} failed.".format(self._name))
-        else:
-            return True
-
-    @property
-    def enabled(self):
-        """Flag indicating if the setup is enabled."""
-        return self.get_sim_setup_info.SimulationSettings.Enabled
-
-    @enabled.setter
-    def enabled(self, value):
-        self.get_sim_setup_info.SimulationSettings.Enabled = value
-        self._edb_object = self._set_edb_setup_info(self.get_sim_setup_info)
-        self._update_setup()
-
-    @property
-    def name(self):
-        """Name of the setup."""
-        return self._edb_object.GetName()
-
-    @name.setter
-    def name(self, value):
-        self._pedb.layout.cell.DeleteSimulationSetup(self.name)
-        edb_setup_info = self.get_sim_setup_info
-        edb_setup_info.Name = value
-        self._name = value
-        self._edb_object = self._set_edb_setup_info(edb_setup_info)
-        self._update_setup()
-
-    @property
-    def position(self):
-        """Position in the setup list."""
-        return self.get_sim_setup_info.Position
-
-    @position.setter
-    def position(self, value):
-        edb_setup_info = self.get_sim_setup_info.SimulationSettings
-        edb_setup_info.Position = value
-        self._set_edb_setup_info(edb_setup_info)
-        self._update_setup()
-
-    @property
-    def setup_type(self):
-        """Type of the setup."""
-        return self.get_sim_setup_info.SimSetupType.ToString()
-
-    @property
-    def frequency_sweeps(self):
-        """List of frequency sweeps."""
-        temp = {}
-        if self.setup_type in ("kRaptorX", "kHFSSPI"):
-            sweep_data_list = self._edb_setup_info.SweepDataList
-        else:
-            sweep_data_list = self.get_sim_setup_info.SweepDataList
-        for i in list(sweep_data_list):
-            temp[i.Name] = EdbFrequencySweep(self, None, i.Name, i)
-        return temp
-
-    @pyedb_function_handler
-    def _add_frequency_sweep(self, sweep_data):
-        """Add a frequency sweep.
-
-        Parameters
-        ----------
-        sweep_data: EdbFrequencySweep
-        """
-        self._sweep_list[sweep_data.name] = sweep_data
-        if self.setup_type in ["kRaptorX", "kHFSSPI"]:
-            edb_setup_info = self._edb_setup_info
-        else:
-            edb_setup_info = self.get_sim_setup_info
-
-        if self._setup_type in ["kSIwave", "kHFSS", "kRaptorX", "kHFSSPI"]:
-            for _, v in self._sweep_list.items():
-                edb_setup_info.SweepDataList.Add(v._edb_object)
-
-        self._edb_object = self._set_edb_setup_info(edb_setup_info)
-        self._update_setup()
-
-    @pyedb_function_handler
-    def delete_frequency_sweep(self, sweep_data):
-        """Delete a frequency sweep.
-
-        Parameters
-        ----------
-            sweep_data : EdbFrequencySweep.
-        """
-        name = sweep_data.name
-        if name in self._sweep_list:
-            self._sweep_list.pop(name)
-
-        fsweep = []
-        if self.frequency_sweeps:
-            fsweep = [val for key, val in self.frequency_sweeps.items() if not key == name]
-            self.get_sim_setup_info.SweepDataList.Clear()
-            for i in fsweep:
-                self.get_sim_setup_info.SweepDataList.Add(i._edb_object)
-            self._update_setup()
-            return True if name in self.frequency_sweeps else False
-
-    @pyedb_function_handler()
-    def add_frequency_sweep(self, name=None, frequency_sweep=None):
-        """Add frequency sweep.
-
-        Parameters
-        ----------
-        name : str, optional
-            Name of the frequency sweep. The default is ``None``.
-        frequency_sweep : list, optional
-            List of frequency points. The default is ``None``.
-
-        Returns
-        -------
-        :class:`pyedb.dotnet.edb_core.edb_data.simulation_setup_data.EdbFrequencySweep`
-
-        Examples
-        --------
-        >>> setup1 = edbapp.create_siwave_syz_setup("setup1")
-        >>> setup1.add_frequency_sweep(frequency_sweep=[
-        ...     ["linear count", "0", "1kHz", 1],
-        ...     ["log scale", "1kHz", "0.1GHz", 10],
-        ...     ["linear scale", "0.1GHz", "10GHz", "0.1GHz"],
-        ...     ])
-        """
-        if name in self.frequency_sweeps:
-            return False
-
-        if not frequency_sweep:
-            frequency_sweep = [["linear scale", "0.1GHz", "10GHz", "0.1GHz"]]
-        elif not isinstance(frequency_sweep[0], list):
-            frequency_sweep = [frequency_sweep]
-
-        if not name:
-            name = generate_unique_name("sweep")
-        sweep = EdbFrequencySweep(self, frequency_sweep, name)
-        self._add_frequency_sweep(sweep)
-        self._update_setup()
-        return sweep
-
-
 class EdbFrequencySweep(object):
     """Manages EDB methods for a frequency sweep.
